requirements.txt
amqp==5.3.1
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
boto3==1.40.9
botocore==1.40.9
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
collectfasta==3.3.0
colorama==0.4.6
coverage==7.10.3
cron-descriptor==1.4.5
cssbeautifier==1.15.4
decorator==5.2.1
django==5.2.5
django-appconf==1.1.0
django-celery-beat==2.8.1
django-celery-email==3.0.0
django-cors-headers==4.7.0
django-coverage-plugin==3.1.1
django-environ==0.12.0
django-extensions==4.1
django-health-check==3.20.0
django-model-utils==5.0.0
django-redis==6.0.0
django-silk==5.4.1
django-storages==1.14.6
django-stubs==5.2.2
django-stubs-ext==5.2.2
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.1
djlint==1.36.4
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
editorconfig==0.17.1
elastic-transport==8.17.1
elasticsearch==8.17.2
executing==2.2.0
factory-boy==3.3.3
faker==37.5.3
flower==2.0.1
googleapis-common-protos==1.70.0
gprof2dot==2025.4.14
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httptools==0.6.4
humanize==4.12.3
idna==3.10
importlib-metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.4.0
ipython-pygments-lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markupsafe==3.0.2
matplotlib-inline==0.1.7
mypy==1.17.1
mypy-extensions==1.1.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-boto3sqs==0.57b0
opentelemetry-instrumentation-botocore==0.57b0
opentelemetry-instrumentation-celery==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-elasticsearch==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-pika==0.57b0
opentelemetry-instrumentation-psycopg==0.57b0
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-propagator-aws-xray==1.0.2
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pika==1.3.2
pillow==11.3.0
pluggy==1.6.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pure-eval==0.2.3
pycparser==2.22
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-sugar==1.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rpds-py==0.27.0
ruff==0.12.8
s3transfer==0.13.1
sentry-sdk==2.34.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
termcolor==3.1.0
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-pyyaml==6.0.12.20250809
types-requests==2.32.4.20250809
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvicorn-worker==0.3.0
vine==5.1.0
wait-for-it==2.3.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
